1.0

+++ ProductionAccounting/ViewModels/WorkDataViewModel.cs
@@ -4,12 +4,11 @@ using Microsoft.EntityFrameworkCore;
 using ProductionAccounting.DAL.Entities;
 using ProductionAccounting.Interfaces;
 using ProductionAccounting.Models;
+using ProductionAccounting.Services.Interfaces;
 using System;
-using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.ComponentModel;
 using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Data;
 using System.Windows.Input;
@@ -18,13 +17,15 @@ namespace ProductionAccounting.ViewModels
 {
     public class WorkDataViewModel : ViewModel
     {
-        private readonly IRepository<Employee> _employeeRepository;
-        private readonly IRepository<Operation> _operationRepository;
         private readonly IRepository<ExecutedOperation> _execRepository;
+        private readonly IConfirmDeleteDialog _confirmDeleteDialog;
+        private readonly IShowExceptionDialogService _showExceptionDialog;
 
-        public WorkDataViewModel(IRepository<ExecutedOperation> execRepository)
+        public WorkDataViewModel(IRepository<ExecutedOperation> execRepository, IConfirmDeleteDialog confirmDeleteDialog, IShowExceptionDialogService showExceptionDialog)
         {
             _execRepository = execRepository;
+            _confirmDeleteDialog = confirmDeleteDialog;
+            _showExceptionDialog = showExceptionDialog;
         }
 
         private bool _onLoading;
@@ -57,16 +58,25 @@ namespace ProductionAccounting.ViewModels
         private async void GetExecutedOperationsViewCommandExecuted()
         {
             OnLoading = true;
-
-            var task = Task.Run(() =>
+            try
             {
-                return _execRepository.Items.ToList().Select(e => new ExecutedOperationDataGridModel(e));
-            });
-            var executedOperations = await task;
-            ExecutedOperation = executedOperations.ToObservableCollection();
-            OperationView = CollectionViewSource.GetDefaultView(ExecutedOperation);
-            OperationView.Filter += _dataView_Filter;
-            OnLoading = false;
+                var task = Task.Run(() =>
+                {
+                    return _execRepository.Items.ToList().Select(e => new ExecutedOperationDataGridModel(e));
+                });
+                var executedOperations = await task;
+                ExecutedOperation = executedOperations.ToObservableCollection();
+                OperationView = CollectionViewSource.GetDefaultView(ExecutedOperation);
+                OperationView.Filter += _dataView_Filter;
+            }
+            catch (System.Exception ex)
+            {
+                _showExceptionDialog.ShowDialog("В работе приложения произошла ошибка. Попробуйте еще раз.\nПоказать сообщения для разработчика?", ex.Message);
+            }
+            finally
+            {
+                OnLoading = false;
+            }
         }
 
         #endregion
@@ -80,19 +90,30 @@ namespace ProductionAccounting.ViewModels
 
         private async void DeleteExecutedOperationsViewCommandExecuted()
         {
-            OnLoading = true;
-            await Task.Run(async () => 
+            try
             {
-                var deleteItem = await _execRepository.Items.FirstOrDefaultAsync(i => i.Id == SelectedItem.Id);
-                if (deleteItem != null)
+                if (!_confirmDeleteDialog.ShowConfirmDeleteDialog("Удалить эту запись?")) return;
+                OnLoading = true;
+                await Task.Run(async () =>
                 {
-                    await _execRepository.DeleteAsync(deleteItem.Id);
-                    await _execRepository.SaveChangesAsync();
-                }
-            });
-            ExecutedOperation.Remove(SelectedItem);
-            SelectedItem = null;
-            OnLoading = false;
+                    var deleteItem = await _execRepository.Items.FirstOrDefaultAsync(i => i.Id == SelectedItem.Id);
+                    if (deleteItem != null)
+                    {
+                        await _execRepository.DeleteAsync(deleteItem.Id);
+                        await _execRepository.SaveChangesAsync();
+                    }
+                });
+                ExecutedOperation.Remove(SelectedItem);
+                SelectedItem = null;
+            }
+            catch (System.Exception ex)
+            {
+                _showExceptionDialog.ShowDialog("В работе приложения произошла ошибка. Попробуйте еще раз.\nПоказать сообщения для разработчика?", ex.Message);
+            }
+            finally
+            {
+                OnLoading = false;
+            }
         }
 
         #endregion
@@ -108,8 +129,8 @@ namespace ProductionAccounting.ViewModels
             }
         }
 
-        private int _monthFilter = DateTime.Now.Month;
-        public int MonthFilter
+        private string _monthFilter = DateTime.Now.Month.ToString();
+        public string MonthFilter
         {
             get { return _monthFilter; }
             set
@@ -119,8 +140,8 @@ namespace ProductionAccounting.ViewModels
             }
         }
 
-        private int _yearFilter = DateTime.Now.Year;
-        public int YearFilter
+        private string _yearFilter = DateTime.Now.Year.ToString();
+        public string YearFilter
         {
             get { return _yearFilter; }
             set
@@ -156,8 +177,8 @@ namespace ProductionAccounting.ViewModels
             if (sender is ExecutedOperationDataGridModel model)
             {
                 return model.Employee.ToString().Contains(NameFilter ?? String.Empty, StringComparison.OrdinalIgnoreCase)
-                && model.Year.ToString().Contains(YearFilter.ToString(), StringComparison.OrdinalIgnoreCase)
-                && model.Month.ToString().Contains(MonthFilter.ToString(), StringComparison.OrdinalIgnoreCase)
+                && model.Year.ToString().Contains(YearFilter ?? String.Empty, StringComparison.OrdinalIgnoreCase)
+                && model.Month.ToString().Contains(MonthFilter ?? String.Empty, StringComparison.OrdinalIgnoreCase)
                 ;
             }
             return false;
